fix(member) 학생 도메인에 사용하는 grade query 분리
